change: use Group in the new_group utils
